# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Visual Diagram Gen API",
    "version": "1.0.0",
    "description": "Backend API for the AI explainer animation tool. This document describes the public HTTP contract used by the frontend and MCP tooling."
  },
  "servers": [
    { "url": "http://localhost:3001", "description": "Local development API" }
  ],
  "security": [{ "bearerAuth": [] }],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": { "type": "string" },
              "message": { "type": "string" },
              "details": { "type": "array", "items": {} }
            },
            "required": ["code", "message", "details"]
          }
        },
        "required": ["error"]
      },
...
      "Project": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "spec": { "type": "object" },
          "maxComponents": { "type": ["integer", "null"], "nullable": true },
          "userId": {
            "type": ["string", "null"],
            "format": "uuid",
            "nullable": true
          },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        },
        "required": ["id", "name", "spec", "createdAt", "updatedAt"]
      },
      "ProjectList": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ProjectSummary" }
          }
        },
        "required": ["items"]
      },
      "ProjectSummary": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        },
        "required": ["id", "name", "createdAt", "updatedAt"]
      },
...
      "Render": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "projectId": { "type": "string", "format": "uuid" },
          "status": {
            "type": "string",
            "enum": ["queued", "running", "done", "failed"]
          },
          "progress": { "type": "integer", "minimum": 0, "maximum": 100 },
          "specSnapshot": { "type": "object" },
          "outputAssetId": {
            "type": ["string", "null"],
            "format": "uuid",
            "nullable": true
          },
          "error": { "type": ["string", "null"], "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        },
        "required": [
          "id",
          "projectId",
          "status",
          "progress",
          "specSnapshot",
          "createdAt",
          "updatedAt"
        ]
      },
      "ValidationResult": {
        "type": "object",
        "properties": {
          "valid": { "type": "boolean" },
          "issues": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Issue" }
          }
        },
        "required": ["valid", "issues"]
      },
      "Issue": {
        "type": "object",
        "properties": {
          "path": { "type": "string" },
          "code": { "type": "string" },
          "message": { "type": "string" }
        },
        "required": ["path", "code", "message"]
      },
      "ProjectCreateRequest": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "minLength": 1 },
          "spec": { "type": "object" }
        },
        "required": ["name"]
      },
      "ProjectUpdateRequest": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "minLength": 1 },
          "spec": { "type": "object" }
        },
        "minProperties": 1
      },
      "ProjectSettingsRequest": {
        "type": "object",
        "properties": {
          "maxComponents": {
            "type": ["integer", "null"],
            "minimum": 1,
            "maximum": 300,
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ItemsRenderResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Render" }
          }
        },
        "required": ["items"]
      },
...
    },
    "parameters": {
      "ProjectId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "string", "format": "uuid" }
      },
...
      "RenderId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "string", "format": "uuid" }
      }
    },
    "responses": {
...
      "NotFound": {
        "description": "Not found",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
      "Conflict": {
        "description": "Conflict",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
      "ValidationFailed": {
        "description": "Semantic validation failed",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
...
      "ProjectCreated": {
        "description": "Project created",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Project" }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/projects": {
      "get": {
        "summary": "List projects",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of projects",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ProjectList" }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new project",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ProjectCreateRequest" }
            }
          }
        },
        "responses": {
          "201": { "$ref": "#/components/responses/ProjectCreated" }
        }
      }
    },
    "/projects/{id}": {
      "get": {
        "summary": "Get a project by id",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/ProjectId" }],
        "responses": {
          "200": {
            "description": "Project record",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Project" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "put": {
        "summary": "Update a project",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/ProjectId" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ProjectUpdateRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated project",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Project" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" },
          "422": { "$ref": "#/components/responses/ValidationFailed" }
        }
      },
      "delete": {
        "summary": "Delete a project",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/ProjectId" }],
        "responses": {
          "204": { "description": "Project deleted" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/projects/{id}/validate": {
      "post": {
        "summary": "Validate a spec without saving it",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/ProjectId" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": { "spec": { "type": "object" } },
                "required": ["spec"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Validation result",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ValidationResult" }
              }
            }
          }
        }
      }
    },
    "/projects/{id}/settings": {
      "patch": {
        "summary": "Set or clear the project component budget",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/ProjectId" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectSettingsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated project",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Project" }
              }
            }
          }
        }
      }
    },
...
    "/projects/{id}/renders": {
      "post": {
        "summary": "Create a render job for a project",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/ProjectId" }],
        "responses": {
          "202": {
            "description": "Render queued",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Render" }
              }
            }
          },
          "422": { "$ref": "#/components/responses/ValidationFailed" }
        }
      },
      "get": {
        "summary": "List render jobs for a project",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/ProjectId" }],
        "responses": {
          "200": {
            "description": "List of renders",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ItemsRenderResponse" }
              }
            }
          }
        }
      }
    },
    "/renders/{id}": {
      "get": {
        "summary": "Get render status by id",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/RenderId" }],
        "responses": {
          "200": {
            "description": "Render status",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Render" }
              }
            }
          }
        }
      }
    },
    "/renders/{id}/file": {
      "get": {
        "summary": "Download a completed render file",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/RenderId" }],
        "responses": {
          "200": {
            "description": "Binary video file",
            "content": {
              "video/mp4": {
                "schema": { "type": "string", "format": "binary" }
              }
            }
          },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    }
  }
}